Removed "highload" as it is un-needed
The script is designed to set 3 CPU frequencies - min, medium & max. Before this change, it would check if the CPU was above "highload" before setting it to max speed. This would mean that the script would be NOOP, between "midload" & "highload" values.
Changed that, so it goes to max, above midload value. 

Other scaling behaviour is unchanged.